third_party/blink/web_tests/flag-specific/highdpi/virtual/text-antialias/font-features/caps-casemapping-expected.txt

This is a testharness.js-based test.
[FAIL] Synthetic small caps should produce the same width as manually uppercased, downscaled text.
  assert_equals: expected 3740.666748046875 but got 3822.666748046875
[FAIL] Strings containing uppercased sharp S should not be truncated.
  assert_equals: expected 68.66667175292969 but got 70
Harness: the test ran to completion.
